# openapi/openapi.yaml
openapi: "3.0.0"
info:
  version: 0.1.5
  title: CRM API
  description: A main CRM Api
  contact:
    name: API Team
servers:
  - url: "localhost:8080"
    description: Dev server
...
paths:
  /legal-entities:
    get:
      tags: [legalentities]
      summary: Get all legal entities
      responses:
        200:
          description: List of legal entities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LegalEntityDTO'
    post:
      tags: [legalentities]
      summary: Create legal entity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LegalEntityDTO'
      responses:
        201:
          description: Legal entity created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LegalEntityDTO'
  /legal-entities/{uuid}:
    put:
      tags: [legalentities]
      summary: Update legal entity
      parameters:
        - in: path
          name: uuid
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LegalEntityDTO'
      responses:
        200:
          description: Legal entity updated
    delete:
      tags: [legalentities]
      summary: Delete legal entity
      parameters:
        - in: path
          name: uuid
          required: true
          schema:
            type: string
      responses:
        204:
          description: Legal entity deleted
...
components:
...
  schemas:
...
    LegalEntityDTO:
      type: object
      properties:
        uuid:
          type: string
        name:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [name]
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []